+++ src/AzerothWarsCSharp.Source/Mechanics/Goblins/Trader.cs
@@ -1,13 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using AzerothWarsCSharp.MacroTools;
 using AzerothWarsCSharp.MacroTools.Buffs;
 using AzerothWarsCSharp.MacroTools.SpellSystem;
 using WCSharp.Buffs;
 using WCSharp.Shared.Data;
 using static War3Api.Common;
 
-namespace AzerothWarsCSharp.MacroTools.UnitEffects
+namespace AzerothWarsCSharp.Source.Mechanics.Goblins
 {
   /// <summary>
   /// When trained, the <see cref="Trader"/> patrols between the originating point and a random trade destination.
