+++ src/AzerothWarsCSharp.Source/Setup/SpellsSetup.cs
@@ -1,6 +1,7 @@
 ﻿using AzerothWarsCSharp.MacroTools.Spells;
 using AzerothWarsCSharp.MacroTools.SpellSystem;
 using AzerothWarsCSharp.MacroTools.UnitEffects;
+using AzerothWarsCSharp.Source.Mechanics.Goblins;
 using AzerothWarsCSharp.Source.Spells;
 using static War3Api.Common;
